feat: add insert functionality to BQ base class #12
