+++ op_tcg/backend/etl/classes.py
@@ -7,7 +7,7 @@ from op_tcg.backend.elo import EloCreator
 from op_tcg.backend.etl.base import AbstractETLJob, E, T
 from op_tcg.backend.etl.load import get_or_create_table
 from op_tcg.backend.etl.transform import BQMatchCreator
-from op_tcg.backend.models.bq import BQDataset
+from op_tcg.backend.models.bq_enums import BQDataset
 from op_tcg.backend.models.input import AllLeaderMetaDocs, MetaFormat, LimitlessLeaderMetaDoc
 from op_tcg.backend.models.matches import BQMatches, Match, BQLeaderElos, LeaderElo
 from op_tcg.backend.etl.extract import read_json_files
