+++ op_tcg/backend/models/base.py
@@ -1,8 +1,8 @@
 import re
 from enum import Enum
-from typing import ClassVar
 
 from pydantic import BaseModel
+
 
 class EnumBase(Enum):
     @classmethod
@@ -10,17 +10,19 @@ class EnumBase(Enum):
         return list(map(lambda c: c.value, cls))
 
 
-
 def classproperty(func):
     class _ClassProperty:
         def __get__(self, instance, owner):
             return func(owner)
+
     return _ClassProperty()
 
 
-class BQTableBaseModel(BaseModel):
+
+class SQLTableBaseModel(BaseModel):
 
     @classproperty
     def __tablename__(cls) -> str:
         # use snake case and not just lower case
         return re.sub(r'(?<!^)(?=[A-Z])', '_', cls.__name__).lower()
+
