+++ op_tcg/backend/models/common.py
@@ -3,3 +3,4 @@ from enum import StrEnum
 
 class DataSource(StrEnum):
     LIMITLESS = "limitless_tcg"
+
